# .context/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: JobWise API
  description: |
    JobWise AI-powered job application assistant API.
    
    ## Features
    - Master resume profile management
    - Job discovery and management
    - AI-powered resume and cover letter generation (5-stage pipeline)
    - Document management and PDF generation
    - Real-time generation progress tracking
    
    ## Authentication
    - Development: API Key authentication
    - Production: JWT Bearer tokens with OAuth 2.0
    
    ## Rate Limiting
    - AI Generation: 10 requests/hour per user
    - Other endpoints: 100 requests/minute per user
    
  version: 1.0.0
  contact:
    name: JobWise Development Team
    email: dev@jobwise.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.jobwise.ai/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /documents:
    get:
      tags:
        - Documents
      summary: List generated documents
      description: Retrieve user's generated documents with filtering
      parameters:
        - name: document_type
          in: query
          description: Filter by document type
          schema:
            type: string
            enum: [resume, cover_letter]
        - name: job_id
          in: query
          description: Filter by associated job
          schema:
            type: string
        - name: profile_id
          in: query
          description: Filter by source profile
          schema:
            type: string
        - name: created_after
          in: query
          description: Filter by creation date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
  /documents/{document_id}:
    get:
      tags:
        - Documents
      summary: Get document details
      description: Retrieve detailed information about generated document
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
              example:
                id: "doc_101"
                document_type: "resume"
                title: "Senior Python Developer - TechCorp Resume"
                content:
                  text: "John Doe\nSoftware Engineer\n\nPROFESSIONAL SUMMARY\n..."
                  html: "<html><body>..."
                metadata:
                  generation_id: "gen_789"
                  profile_id: "profile_456"
                  job_id: "job_123"
                  job_title: "Senior Python Developer"
                  company: "TechCorp Inc"
                  template: "modern"
                  ats_score: 0.87
                  match_percentage: 82
                  keyword_coverage: 0.91
                  tokens_used: 7850
                  generation_time: 28.5
                pdf:
                  url: "https://api.jobwise.ai/documents/doc_101/download"
                  size_bytes: 245678
                  page_count: 1
                version: 1
                created_at: "2024-10-18T10:31:28Z"
                updated_at: "2024-10-18T10:31:28Z"

    delete:
      tags:
        - Documents
      summary: Delete document
      description: Permanently delete generated document and PDF
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204':
          description: Document deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /documents/{document_id}/download:
    get:
      tags:
        - Documents
      summary: Download document PDF
      description: Download generated document as PDF file
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
                example: "attachment; filename=\"John_Doe_Resume_TechCorp.pdf\""
            Content-Length:
              description: File size in bytes
              schema:
                type: integer
        '404':
          $ref: '#/components/responses/NotFound'

  # POST /documents is handled above at /documents path
    
    DocumentId:
      name: document_id
      in: path
      required: true
      description: Document unique identifier
      schema:
        type: string
        format: uuid
...